Do not install the captain in CI
This modification will not install The Cap'n in CI-Environments (at least those that go with the de-facto standard to set the `CI` environment variable to `true`).

Installing The Cap'n in such environments does not make any sense as there are either no commits going to happen or if so they are going to happen in an automated fashion so breaking those commits is not really helpful. And having  a CI-pipeline run by The Cap'n inside a CI-environment feels... weird to say the least.